Updated plotting structure. Fixed minor bugs.
